fix: check for component before destroying context in ProjectPreview

+++ packages/haiku-creator/src/react/components/ProjectPreview.js
@@ -51,7 +51,9 @@ class ProjectPreview extends React.Component {
 
   componentWillUnmount () {
     this.stopComponentClock() // Avoid wasted CPU rendering for unseen DOM nodes
-    this.component.context.destroy()
+    if (this.component) {
+      this.component.context.destroy()
+    }
   }
 
   componentDidMount () {
